requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.5
dictdiffer==0.9.0
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
pydantic==2.0.3
pydantic_core==2.3.0
pypika-tortoise==0.1.6
python-multipart==0.0.6
pytz==2023.3
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.11.8
tortoise-orm==0.19.3
typing_extensions==4.7.1
uvicorn==0.23.1
